Refactor the audio player into its own class.
