+++ Ghidra/Framework/SoftwareModeling/src/main/java/ghidra/program/model/data/WAVEDataType.java
@@ -15,18 +15,10 @@
  */
 package ghidra.program.model.data;
 
-import java.awt.event.MouseEvent;
-import java.io.ByteArrayInputStream;
-import java.io.IOException;
-
-import javax.sound.sampled.*;
-import javax.swing.ImageIcon;
-
 import ghidra.docking.settings.Settings;
 import ghidra.program.model.mem.MemBuffer;
 import ghidra.program.model.mem.MemoryAccessException;
 import ghidra.util.Msg;
-import resources.ResourceManager;
 
 public class WAVEDataType extends BuiltIn implements Dynamic {
 	public static byte[] MAGIC = new byte[] { (byte) 'R', (byte) 'I', (byte) 'F', (byte) 'F',
@@ -101,37 +93,6 @@ public class WAVEDataType extends BuiltIn implements Dynamic {
 		return "<WAVE-Resource>";
 	}
 
-	private static class WAVEData implements Playable {
-
-		private static final ImageIcon AUDIO_ICON =
-			ResourceManager.loadImage("images/audio-volume-medium.png");
-		private byte[] bytes;
-
-		public WAVEData(byte[] bytes) {
-			this.bytes = bytes;
-		}
-
-		@Override
-		public void clicked(MouseEvent event) {
-
-			try {
-				Clip clip = AudioSystem.getClip();
-				AudioInputStream ais =
-					AudioSystem.getAudioInputStream(new ByteArrayInputStream(bytes));
-				clip.open(ais);
-				clip.start();
-			}
-			catch (UnsupportedAudioFileException | IOException | LineUnavailableException e) {
-				Msg.debug(this, "Unable to play audio", e);
-			}
-		}
-
-		@Override
-		public ImageIcon getImageIcon() {
-			return AUDIO_ICON;
-		}
-	}
-
 	@Override
 	public Object getValue(MemBuffer buf, Settings settings, int length) {
 		byte[] data = new byte[length];
@@ -139,13 +100,12 @@ public class WAVEDataType extends BuiltIn implements Dynamic {
 			Msg.error(this, "WAVE-Sound error: " + "Not enough bytes!");
 			return null;
 		}
-		WAVEData waveData = new WAVEData(data);
-		return waveData;
+		return new DefaultAudioPlayer(data);
 	}
 
 	@Override
 	public Class<?> getValueClass(Settings settings) {
-		return WAVEData.class;
+		return DefaultAudioPlayer.class;
 	}
 
 	@Override
